reply완성

+++ src/main/java/com/example/clone6_backend/controller/ReplyController.java
@@ -2,6 +2,8 @@ package com.example.clone6_backend.controller;
 
 import com.example.clone6_backend.dto.request.ReplyRequestDto;
 import com.example.clone6_backend.dto.response.ReplyResponseDto;
+import com.example.clone6_backend.exceptionHandler.CustomException;
+import com.example.clone6_backend.exceptionHandler.ErrorCode;
 import com.example.clone6_backend.security.UserDetailsImpl;
 import com.example.clone6_backend.service.ReplyService;
 import lombok.RequiredArgsConstructor;
@@ -17,16 +19,20 @@ public class ReplyController {
 
     @PostMapping("/api/fund/comment/{commentId}/reply")
     public ReplyResponseDto createReply(@RequestBody ReplyRequestDto requestDto, @PathVariable Long commentId,@AuthenticationPrincipal UserDetailsImpl userDetails){
+        if(userDetails.equals("")){throw new CustomException(ErrorCode.USER_NOT_FOUND);}
+
         return replyService.createReply(requestDto,commentId,userDetails);
     }
 
     @PutMapping("/api/fund/commment/{commentId}/reply")
     public ReplyResponseDto update(@RequestBody ReplyRequestDto requestDto, @PathVariable Long commentId,@AuthenticationPrincipal UserDetailsImpl userDetails){
+        if(userDetails.equals("")){throw new CustomException(ErrorCode.USER_NOT_FOUND);}
         return replyService.update(requestDto, commentId, userDetails);
     }
 
     @DeleteMapping("/api/fund/comment/reply/{replyId}")
     public ResponseEntity delete(@PathVariable Long replyId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
-        return replyService.delete(replyId, userDetails);
+            if(userDetails.equals("")){throw new CustomException(ErrorCode.USER_NOT_FOUND);}
+            return replyService.delete(replyId, userDetails);
     }
 }
